+++ luminary/Providers/LuminaryServiceProvider.php
@@ -4,6 +4,7 @@ namespace Luminary\Providers;
 
 use Illuminate\Support\ServiceProvider;
 use Luminary\Services\ApiQuery\ServiceProvider as ApiQueryServiceProvider;
+use Luminary\Services\Testing\ServiceProvider as TestingServiceProvider;
 use Luminary\Services\Filesystem\FilesystemServiceProvider;
 use Luminary\Services\FormRequests\ServiceProvider as FormReqeustsServiceProvider;
 
@@ -19,6 +20,7 @@ class LuminaryServiceProvider extends ServiceProvider
         FilesystemServiceProvider::class,
         FormReqeustsServiceProvider::class,
         ApiQueryServiceProvider::class,
+        TestingServiceProvider::class
     ];
 
     /**
