+++ luminary/Services/ApiQuery/ServiceProvider.php
@@ -3,7 +3,6 @@
 namespace Luminary\Services\ApiQuery;
 
 use Illuminate\Support\ServiceProvider as LaravelServiceProvider;
-use Illuminate\Database\Eloquent\Factory as ModelFactory;
 
 class ServiceProvider extends LaravelServiceProvider
 {
@@ -26,28 +25,5 @@ class ServiceProvider extends LaravelServiceProvider
         });
 
         $this->app->middleware(QueryMiddleware::class);
-
-        if (app()->environment() == 'testing') {
-            $this->registerTesting();
-        }
-    }
-
-    /**
-     * Register the testing database migrations
-     * and factories
-     *
-     * @return void
-     */
-    public function registerTesting()
-    {
-        $dir = __DIR__ . '/Testing/database';
-
-        // Register Factory
-        app(ModelFactory::class)->load($dir . '/factories');
-
-        // Register Migrations
-        $this->app->afterResolving('migrator', function ($migrator) use ($dir) {
-            $migrator->path($dir . '/migrations');
-        });
     }
 }
