+++ tests/ApiQuery/QueryTest.php
@@ -4,7 +4,7 @@ use Luminary\Services\ApiQuery\QueryCollection;
 
 class QueryTest extends TestCase
 {
-    use Luminary\Services\ApiQuery\Testing\BaseQueryTrait;
+    use Luminary\Services\Testing\BaseQueryTrait;
 
     /**
      * Query String parameters to build
